requirements.txt
replicate
python-dotenv
requests
Pillow
colorama
google-generativeai
